PTDLvHS/requirements.txt
torch
torchvision
timm
numpy
matplotlib
pillow
flask